# Heapsnapshot exporting optimization.