--- a/andb/loader/__init__.py
+++ b/andb/loader/__init__.py
@@ -1,3 +1,3 @@
 from .elf import *
 from .core import *
-
+from .loader import *