--- a/init/pre_mapreduce.cmd
+++ b/init/pre_mapreduce.cmd
@@ -1,0 +1,3 @@
+iso g p 
+andb option chunk_cache on
+