Hi, All

for large corefile, the core.heapsnapshot exporting costs lots of time. 

the optimization introduced a new command, 

```
andb -c core.33 -m snapshot -j 8
```
to export core.heapsnapshot by using N processes concurrently.
it can reduce N times real time if you have a multi-cores CPU installed.

In my test, core.33 is about 1.8G bytes and has about 8m nodes and 32m edges,
normally costs 3+hrs for a single process exporting, 
now only costs 7min on my M1 MBP.
```
...
Polling done.
Resolve all edges ...
Done, resolved 4162, failed 0.
Total Entry(7782138), Edge(31256137)
heap snapshot written to 'core.heapsnapshot'
ReduceGenerate() takes 416.855 second(s).
real   435.652s
map    7.047s
parse  358.794s
reduce 69.811s
```

Get the object's memory address in Heapsnapshot,
another optimization is andb saved the memory address for each HeapObject in core.heapsnapshot file, let you inspect more in the debugger. each HeapObject has a unique Index normally can be identified from Devtools GUI, the number with '@' prefix after an object name.

`grep` it from core.heapsnapshot, memory address sits in the last.
```
$ grep ,6173663, core.heapsnapshot
,3,766,6173663,48,5,0,18796081870720

# inspect in andb
(lldb) p/x 18796081870720
(long) $0 = 0x000011184dc87780
(lldb) v8 o 0x000011184dc87780
[JSObject 0x11184dc87781]
- Properties:
 - _Logger: <JsObject 0x11184dc886a9>
 - _mod: <JsObject 0x11184dc886e1>
 - ErrorInfo: <JsArray 0x11184dc886f9>
- Elements: []
 - constructor: <JsFunction AAALoader 0x2e8c174333b9>
 - name: AAALoader
[Map 0x2342b62e0b9]
- InstanceSizeInWords: 6
- InobjectPropertiesStartOrConstructorFunctionIndex: 3
- UsedOrUnusedInstanceSizeInWords: 6
- VisitorId: 27
- InstanceType: JS_OBJECT_TYPE (1057)
- BitField: 0x0
...
```

some of the python files had been renamed, may need *.pyc cleaning up if odds.

Regards,
Zhaolei